Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    float num;
-    scanf("%f",&num);
-    printf("You entered: %.4f",num);
+    double num;
+    scanf("%lf",&num);
+    printf("You entered: %lf",num);
     return 0;
 }
